pyproject.toml
[tool.poetry]
name = "subnet_ping_api"
version = "0.1.0"
description = ""
authors = ["emrekndl <kondulemre@gmail.com>"]
readme = "README.md"


packages = [
    { include = "subnet_ip" },
    { include = "subnet_ip_app" }
]

[tool.poetry.dependencies]
python = "3.11.0"
amqp = "5.1.1"
celery = "5.2.7"
django = "4.0.8"
django-environ = "0.9.0"
django-extensions = "3.1.5"
django-redis = "5.2.0"
djangorestframework = "3.13.1"
drf-yasg = "1.20.3"
netaddr = "0.8.0"
redis = "4.3.4"
requests = "2.28.1"
python-dotenv = "0.21.0"
dj-database-url = "2.2.0"
gunicorn = "20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
